refactor(sdk): own ready withdrawal selection
