+++ packages/sdk/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./sdk.js";
 export * from "./constants.js";
 export {
-  selectExactReadyWithdrawalDeposits,
   selectReadyWithdrawalCleanupDeposit,
   selectReadyWithdrawalDeposits,
 } from "./withdrawal_selection.js";
