+++ packages/sdk/src/withdrawal_selection.ts
@@ -1,6 +1,6 @@
 import { ccc } from "@ckb-ccc/core";
 import { type IckbDepositCell } from "@ickb/core";
-import { compareBigInt, selectBoundedUdtSubset } from "@ickb/utils";
+import { compareBigInt } from "@ickb/utils";
 
 const READY_POOL_BUCKET_SPAN_MS = 15n * 60n * 1000n;
 const NEAR_READY_LOOKAHEAD_MS = 60n * 60n * 1000n;
@@ -145,13 +145,7 @@ function selectReadyWithdrawalDepositsWithScore(
   );
 }
 
-export function selectExactReadyWithdrawalDeposits(
-  options: ExactReadyWithdrawalSelectionOptions,
-): ReadyWithdrawalSelection | undefined {
-  return selectExactReadyWithdrawalDepositsWithScore(options);
-}
-
-function selectExactReadyWithdrawalDepositsWithScore(
+function selectReadyWithdrawalCandidateWithScore(
   options: ScoredExactReadyWithdrawalSelectionOptions,
 ): ReadyWithdrawalSelection | undefined {
   const { count, ...selectionOptions } = options;
@@ -200,11 +194,11 @@ export function selectExactReadyWithdrawalDepositCandidates(
       count: options.count,
       preserveSingletons: options.preserveSingletons,
     };
-    addSelection(selectExactReadyWithdrawalDepositsWithScore({
+    addSelection(selectReadyWithdrawalCandidateWithScore({
       ...baseOptions,
       score: options.score,
     }));
-    addSelection(selectExactReadyWithdrawalDepositsWithScore(baseOptions));
+    addSelection(selectReadyWithdrawalCandidateWithScore(baseOptions));
   }
 
   return selections;
@@ -255,7 +249,7 @@ function selectReadyDeposits<T extends { udtValue: bigint }>(
     return [];
   }
 
-  const bestFit = selectBoundedUdtSubset(deposits, maxAmount, {
+  const bestFit = selectBoundedReadyDepositSubset(deposits, maxAmount, {
     candidateLimit: BEST_FIT_SEARCH_CANDIDATES,
     minCount: requiredCount,
     maxCount,
@@ -270,6 +264,250 @@ function selectReadyDeposits<T extends { udtValue: bigint }>(
   );
 
   return pickBetterSelection(deposits, bestFit, greedy, score);
+}
+
+function selectBoundedReadyDepositSubset<T extends { udtValue: bigint }>(
+  items: readonly T[],
+  maxAmount: bigint,
+  options: {
+    candidateLimit: number;
+    minCount: number;
+    maxCount: number;
+    score?: (item: T) => bigint;
+  },
+): T[] {
+  const { candidateLimit, minCount, maxCount } = options;
+  const scoreOf = options.score ?? ((item: T): bigint => item.udtValue);
+  const boundedItems = items.slice(0, candidateLimit);
+  const effectiveMaxCount = Math.min(maxCount, boundedItems.length);
+  if (
+    maxAmount <= 0n ||
+    minCount < 0 ||
+    effectiveMaxCount < minCount ||
+    boundedItems.length === 0
+  ) {
+    return [];
+  }
+
+  interface PartialSelection {
+    mask: number;
+    total: bigint;
+    score: bigint;
+  }
+
+  const split = Math.floor(boundedItems.length / 2);
+  const firstHalf = boundedItems.slice(0, split);
+  const secondHalf = boundedItems.slice(split);
+  assertBitmaskSearchSize(firstHalf.length);
+  assertBitmaskSearchSize(secondHalf.length);
+
+  const enumerate = (half: readonly T[]): PartialSelection[][] => {
+    const groups = Array.from(
+      { length: half.length + 1 },
+      () => [] as PartialSelection[],
+    );
+
+    const search = (
+      index: number,
+      mask: number,
+      count: number,
+      total: bigint,
+      score: bigint,
+    ): void => {
+      if (index === half.length) {
+        groups[count]?.push({ mask, total, score });
+        return;
+      }
+
+      search(index + 1, mask, count, total, score);
+
+      const item = half[index];
+      if (item === undefined) {
+        return;
+      }
+      search(
+        index + 1,
+        mask | (1 << index),
+        count + 1,
+        total + item.udtValue,
+        score + scoreOf(item),
+      );
+    };
+
+    search(0, 0, 0, 0n, 0n);
+    return groups;
+  };
+
+  const firstByCount = enumerate(firstHalf);
+  const secondByCount = enumerate(secondHalf).map((selections) =>
+    prepareSelections(selections, secondHalf.length)
+  );
+
+  let best:
+    | {
+        firstMask: number;
+        secondMask: number;
+        total: bigint;
+        score: bigint;
+      }
+    | undefined;
+
+  for (let firstCount = 0; firstCount <= effectiveMaxCount; firstCount += 1) {
+    const firstSelections = firstByCount[firstCount] ?? [];
+    for (const first of firstSelections) {
+      if (first.total > maxAmount) {
+        continue;
+      }
+
+      const minSecondCount = Math.max(0, minCount - firstCount);
+      const maxSecondCount = effectiveMaxCount - firstCount;
+      for (let secondCount = minSecondCount; secondCount <= maxSecondCount; secondCount += 1) {
+        const secondSelections = secondByCount[secondCount] ?? [];
+        const second = findBestAtOrBelow(secondSelections, maxAmount - first.total);
+        if (!second) {
+          continue;
+        }
+
+        const total = first.total + second.total;
+        const candidate = {
+          firstMask: first.mask,
+          secondMask: second.mask,
+          total,
+          score: first.score + second.score,
+        };
+        if (!best || isBetterSelection(candidate, best, firstHalf.length, secondHalf.length)) {
+          best = candidate;
+        }
+      }
+    }
+  }
+
+  if (!best) {
+    return [];
+  }
+
+  return selectByMasks(firstHalf, best.firstMask).concat(
+    selectByMasks(secondHalf, best.secondMask),
+  );
+}
+
+function assertBitmaskSearchSize(length: number): void {
+  if (length > 16) {
+    throw new Error("Bounded subset search supports at most 16 items per half");
+  }
+}
+
+function prepareSelections(
+  selections: { mask: number; total: bigint; score: bigint }[],
+  length: number,
+): { total: bigint; selection: { mask: number; total: bigint; score: bigint } }[] {
+  selections.sort((left, right) => {
+    const totalCompare = compareBigInt(left.total, right.total);
+    if (totalCompare !== 0) {
+      return totalCompare;
+    }
+
+    return compareMask(left.mask, right.mask, length);
+  });
+
+  const prepared: { total: bigint; selection: { mask: number; total: bigint; score: bigint } }[] = [];
+  let best: { mask: number; total: bigint; score: bigint } | undefined;
+  for (const selection of selections) {
+    if (!best || isBetterPartialSelection(selection, best, length)) {
+      best = selection;
+    }
+    prepared.push({ total: selection.total, selection: best });
+  }
+
+  return prepared;
+}
+
+function findBestAtOrBelow(
+  items: readonly { total: bigint; selection: { mask: number; total: bigint; score: bigint } }[],
+  limit: bigint,
+): { mask: number; total: bigint; score: bigint } | undefined {
+  let low = 0;
+  let high = items.length - 1;
+  let bestIndex = -1;
+
+  while (low <= high) {
+    const mid = Math.floor((low + high) / 2);
+    const item = items[mid];
+    if (item === undefined) {
+      break;
+    }
+
+    if (item.total <= limit) {
+      bestIndex = mid;
+      low = mid + 1;
+    } else {
+      high = mid - 1;
+    }
+  }
+
+  return bestIndex >= 0 ? items[bestIndex]?.selection : undefined;
+}
+
+function isBetterPartialSelection(
+  left: { mask: number; total: bigint; score: bigint },
+  right: { mask: number; total: bigint; score: bigint },
+  length: number,
+): boolean {
+  if (left.score !== right.score) {
+    return left.score > right.score;
+  }
+
+  if (left.total !== right.total) {
+    return left.total > right.total;
+  }
+
+  return compareMask(left.mask, right.mask, length) < 0;
+}
+
+function isBetterSelection(
+  left: { firstMask: number; secondMask: number; total: bigint; score: bigint },
+  right: { firstMask: number; secondMask: number; total: bigint; score: bigint },
+  firstLength: number,
+  secondLength: number,
+): boolean {
+  if (left.score !== right.score) {
+    return left.score > right.score;
+  }
+
+  if (left.total !== right.total) {
+    return left.total > right.total;
+  }
+
+  const firstCompare = compareMask(left.firstMask, right.firstMask, firstLength);
+  return firstCompare < 0 ||
+    (firstCompare === 0 && compareMask(left.secondMask, right.secondMask, secondLength) < 0);
+}
+
+function selectByMasks<T>(items: readonly T[], mask: number): T[] {
+  const selected: T[] = [];
+  for (let i = 0; i < items.length; i += 1) {
+    if ((mask & (1 << i)) !== 0) {
+      const item = items[i];
+      if (item !== undefined) {
+        selected.push(item);
+      }
+    }
+  }
+  return selected;
+}
+
+function compareMask(left: number, right: number, length: number): number {
+  for (let i = 0; i < length; i += 1) {
+    const leftHas = (left & (1 << i)) !== 0;
+    const rightHas = (right & (1 << i)) !== 0;
+    if (leftHas === rightHas) {
+      continue;
+    }
+
+    return leftHas ? -1 : 1;
+  }
+
+  return 0;
 }
 
 function classifyReadyDeposits(
